version 1.2.1 make red and pink more recognizable, fix score to be float instead of int

+++ Assignment 2/BubblePopGame/BubblePopGame/PlayerScore.swift
@@ -10,5 +10,5 @@ import Foundation
 struct PlayerScore: Identifiable, Codable {
     var id = UUID()
     let playerName: String
-    var score: Int
+    var score: Float
 }
